Allow automation tests to override modifiers on OpenGL system window
The automation tests bypass the native platform eventing to synthesize
keyboard events. However, not all operations in MatterControl look directly
at the keyboard state but instead rely on the ModifierKeys property of the
system window. Therefore, the automation tests must both set the keyboard
state and the modifier keys on the system window.